+++ SwiftValidators/ValueProvider.swift
@@ -1,0 +1,13 @@
+//
+//  ValueProvider.swift
+//  SwiftValidators
+//
+//  Created by Γιώργος Καϊμακάς on 02/06/16.
+//  Copyright © 2016 Γιώργος Καϊμακάς. All rights reserved.
+//
+
+import Foundation
+
+public protocol ValueProvider {
+	var value: String { get }
+}